refactor: split `utils.mjs`, move `validatePath` & `validateProps` to `validation.mjs`
rename `validateFuncProps` to `validateProps`
