+++ src/commands/cat.mjs
@@ -1,7 +1,7 @@
 import { readFile } from 'node:fs/promises';
 import { join } from 'node:path';
 
-import { validatePath } from '../utils.mjs';
+import { validatePath } from '../utils/validation.mjs';
 
 export const cat = async ({ passedProps, currentWorkingDirectory }) => {
   const pathToFile = passedProps[0];
