+++ src/commands/rm.mjs
@@ -1,7 +1,7 @@
 import { rm as fsRm } from 'node:fs/promises';
 import { join } from 'node:path';
 
-import { validatePath } from '../utils.mjs';
+import { validatePath } from '../utils/validation.mjs';
 
 export const rm = async ({ passedProps, currentWorkingDirectory }) => {
   const pathToFile = passedProps[0];
