+++ src/commands/rn.mjs
@@ -1,7 +1,7 @@
 import { rename as fsRename } from 'node:fs/promises';
 import { join } from 'node:path';
 
-import { validatePath } from '../utils.mjs';
+import { validatePath } from '../utils/validation.mjs';
 
 export const rn = async ({ passedProps, currentWorkingDirectory }) => {
   const pathToFile = passedProps[0];
